Initial common code tweak
