+++ otj-metrics-core/pom.xml
@@ -39,7 +39,10 @@
             <groupId>com.opentable.components</groupId>
             <artifactId>otj-jvm</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>com.fasterxml.jackson.core</groupId>
+            <artifactId>jackson-annotations</artifactId>
+        </dependency>
         <dependency>
             <groupId>com.opentable.components</groupId>
             <artifactId>otj-spring</artifactId>
