+++ otj-metrics-jaxrs/src/main/java/com/opentable/metrics/jaxrs/HealthResource.java
@@ -28,13 +28,11 @@ import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
 import com.codahale.metrics.health.HealthCheck.Result;
-import com.fasterxml.jackson.annotation.JsonValue;
-import com.google.common.collect.ComparisonChain;
 
 import org.apache.commons.lang3.ClassUtils;
-import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.tuple.Pair;
 
+import com.opentable.metrics.SortedEntry;
 import com.opentable.metrics.http.CheckState;
 import com.opentable.metrics.http.HealthController;
 
@@ -72,52 +70,13 @@ public class HealthResource {
         });
 
         if (!all && !rendered.isEmpty()) {
-            Result worstResult = rendered.keySet().iterator().next().result;
+            Result worstResult = rendered.keySet().iterator().next().getResult();
             if (!worstResult.isHealthy()) {
-                rendered.keySet().removeIf(e -> e.result.isHealthy());
+                rendered.keySet().removeIf(e -> e.getResult().isHealthy());
             }
         }
 
         return rendered;
     }
 
-    public static class SortedEntry implements Comparable<SortedEntry> {
-        private final String name;
-        final Result result;
-
-        SortedEntry(String name, Result result) {
-            this.name = name;
-            this.result = result;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public int compareTo(SortedEntry o) {
-            return ComparisonChain.start()
-                    // severity descending
-                    .compare(o.result, result, HealthController::compare)
-                    // name ascending
-                    .compare(getName(), o.getName())
-                    .result();
-        }
-
-        @Override
-        public int hashCode() {
-            return getName().hashCode();
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            return EqualsBuilder.reflectionEquals(this, obj, false);
-        }
-
-        @Override
-        @JsonValue
-        public String toString() {
-            return getName();
-        }
-    }
 }
